avoid recreating the same GPU resource when calling AbstractResource::upload() and check that the resource was actually allocated before destroying it in AbstractResource::dispose()
Former-commit-id: d1f18e51b7af379c7c041ed2fd62b98ba488b83f
Former-commit-id: f6c888745c11d57454e28536281549ea6d69f4ce

+++ framework/src/minko/render/Texture.cpp
@@ -64,13 +64,18 @@ Texture::data(unsigned char* data, DataFormat format)
 void
 Texture::upload()
 {
-	_id = _context->createTexture(_width, _height, false, _optimizeForRenderToTexture);
-	_context->uploadTextureData(_id, _width, _height, 0, &_data[0]);
+    if (_id == -1)
+    	_id = _context->createTexture(_width, _height, false, _optimizeForRenderToTexture);
+	
+    _context->uploadTextureData(_id, _width, _height, 0, &_data[0]);
 }
 
 void
 Texture::dispose()
 {
-	_context->deleteTexture(_id);
-	_id = -1;
+    if (_id != -1)
+    {
+	    _context->deleteTexture(_id);
+	    _id = -1;
+    }
 }
